+++ bindings/web/src/builtin/RandomCapability.ts
@@ -1,0 +1,7 @@
+import { Capability } from "..";
+
+export class RandomCapability implements Capability {
+    generate(dest: Uint8Array, id: number): void {
+        window.crypto.getRandomValues(dest);
+    }
+}