+++ bindings/web/src/builtin/index.ts
@@ -1,0 +1,4 @@
+export { TensorFlowModel } from "./TensorFlowModel";
+export { RandomCapability } from "./RandomCapability";
+
+export const TensorFlowLiteMimeType = "application/tflite-model";